+++ src/remotes/zxsimulator/zxports.ts
@@ -14,6 +14,9 @@ export class ZxPorts {
 	// This map maps port addresses to functions that are executed on a port write.
 	protected outPortMap: Map<number, (port: number, value: number) => void>;
 
+	// This map maps port addresses to functions that are executed on a port read.
+	protected inPortMap: Map<number, (port: number) => number>;
+
 	// The bitmask for the port. Only 1 bits are used to decode.
 	// E.g. the ZX128 does not use bit 1 and bit 15.
 	//public portBitMask=0xFFFF;
@@ -22,12 +25,13 @@ export class ZxPorts {
 	/// Constructor.
 	constructor() {
 		this.ports=new Uint8Array(0x10000);
-		this.outPortMap =new Map<number, (port: number, value: number) => void>();
+		this.outPortMap=new Map<number, (port: number, value: number) => void>();
+		this.inPortMap=new Map<number, (port: number) => number>();
 	}
 
 
 	/**
-	 * Registers a function for a specific port address.
+	 * Registers a function for a write to a specific port address.
 	 * @param port The port address
 	 * @param func The function to execute if the port is written.
 	 */
@@ -36,6 +40,19 @@ export class ZxPorts {
 			this.outPortMap.set(port, func);
 		else
 			this.outPortMap.delete(port);
+	}
+
+
+	/**
+	 * Registers a function for a read to a specific port address.
+	 * @param port The port address
+	 * @param func The function to execute if the port is read.
+	 */
+	public registerInPortFunction(port: number, func: ((port: number) => number)|undefined) {
+		if (func)
+			this.inPortMap.set(port, func);
+		else
+			this.inPortMap.delete(port);
 	}
 
 
@@ -78,7 +95,10 @@ export class ZxPorts {
 	 *  Read 1 byte. Used by the CPU.
 	 */
 	public read(port: number): number {
-		assert(port>=0&&port<0x10000);
+		const func=this.inPortMap.get(port);
+		if (func)
+			return func(port);
+		// If no handling function is registered
 		const value=this.ports[port];
 		return value;
 	}
@@ -89,9 +109,15 @@ export class ZxPorts {
 	 * Executes a custom method.
 	 */
 	public write(port: number, data: number) {
-		const func=this.outPortMap.get(port);
-		if (func)
-			func(port, data);
+		const writefunc=this.outPortMap.get(port);
+		if (writefunc)
+			writefunc(port, data);
+		// Check if there is a handler for the read function implemented.
+		const readFunc=this.inPortMap.get(port);
+		if (!readFunc) {
+			// If not store the data for later reading.
+			this.ports[port]=data;
+		}
 	}